Ran clang-tidy before branching 9.0.0
1. Changed the m4 macros so clang-tidy would run in the plugins directory
2. Added options:
     google-readability-braces-around-statements
     google-readability-casting
     modernize-use-using
     performance-unnecessary-copy-initialization
3. Removed option modernize-use-default-member-init because clang-tidy was
core dumping on TextView and maybe other files

+++ plugins/experimental/ssl_session_reuse/src/ssl_init.cc
@@ -61,7 +61,7 @@ init_ssl_params(const std::string &conf)
   if (ssl_param.key_update_interval > STEK_MAX_LIFETIME) {
     ssl_param.key_update_interval = STEK_MAX_LIFETIME;
     TSDebug(PLUGIN, "KeyUpdateInterval too high, resetting session ticket key rotation to %d seconds",
-            (int)ssl_param.key_update_interval);
+            ssl_param.key_update_interval);
   }
 
   TSDebug(PLUGIN, "init_ssl_params: I %s been configured to initially be stek_master",
